refactor import pattern for sub-modules
