+++ README.md
@@ -64,7 +64,8 @@ modifications. For example, you could create a new trainer from ``SFTTrainer``
 that uses a different loss function:
 
 ```python
-from arctic_training.trainers import SFTTrainer, register
+from arctic_training import register
+from arctic_training.trainer import SFTTrainer
 
 @register
 class CustomTrainer(SFTTrainer):
