+++ arctic_training/callback/__init__.py
@@ -14,3 +14,4 @@
 # limitations under the License.
 
 from .callback import Callback
+from .mixin import CallbackMixin
