+++ arctic_training/config/__init__.py
@@ -14,3 +14,10 @@
 # limitations under the License.
 
 from .base import BaseConfig
+from .checkpoint import CheckpointConfig
+from .data import DataConfig
+from .model import ModelConfig
+from .optimizer import OptimizerConfig
+from .scheduler import SchedulerConfig
+from .tokenizer import TokenizerConfig
+from .trainer import TrainerConfig
