+++ arctic_training/data/__init__.py
@@ -15,4 +15,5 @@
 
 from . import sft_source
 from .factory import DataFactory
+from .sft_factory import SFTDataFactory
 from .source import DataSource
