+++ arctic_training/trainer/__init__.py
@@ -13,4 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from arctic_training.trainer.trainer import Trainer
+from .sft_trainer import SFTTrainer
+from .trainer import Trainer
