+++ projects/mlp_speculator/llama-8b.yaml
@@ -34,9 +34,9 @@ logger:
   file_output_ranks: [0]
 scheduler:
   name: cosine
-  lr: 1e-3
   warmup_ratio: 0.05
 optimizer:
+  lr: 1e-3
   betas: [0.9,0.999]
   weight_decay: 0.1
 checkpoint:
