+++ projects/swiftkv/llama-8b.yaml
@@ -22,13 +22,13 @@ data:
   num_proc: 16
   max_length: 8192
 logger:
-  level: INFO
+  level: WARNING
   output_dir: "./"
   file_output_ranks: [0]
 scheduler:
-  lr: 0.0002
   warmup_ratio: 0.05
 optimizer:
+  lr: 0.0002
   betas: [0.9,0.999]
   weight_decay: 0.0
 exit_iteration: 11
